Fixed issue with not all data being returned
The problem was that the code wasn't reading all of the data in the
chunk. Added a third parameter to receive method to signify whether the
lengths had to match exactly. If true, which is set when reading a chunk
header or chunk, then keep reading till you get it all.